fix: 遵守游戏模式模块分层
